+++ code/search/FullTextSearch.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace SilverStripe\FullTextSearch\Search;
+
 use SilverStripe\Core\Config\Config;
 use SilverStripe\Core\ClassInfo;
 /**
