+++ code/search/SearchIndex.php
@@ -1,5 +1,6 @@
 <?php
-namespace SilverStripe\FullTextSearch;
+
+namespace SilverStripe\FullTextSearch\Search;
 
 use SilverStripe\View\ViewableData;
 use SilverStripe\ORM\DataObject;
@@ -604,96 +605,4 @@ abstract class SearchIndex extends ViewableData
      * to be run before _and/or_ after this.
      */
     abstract public function init();
-}
-
-/**
- * A search index that does nothing. Useful for testing
- */
-abstract class SearchIndex_Null extends SearchIndex
-{
-    public function add($object)
-    {
-    }
-
-    public function delete($base, $id, $state)
-    {
-    }
-
-    public function commit()
-    {
-    }
-}
-
-/**
- * A search index that just records actions. Useful for testing
- */
-abstract class SearchIndex_Recording extends SearchIndex
-{
-    public $added = array();
-    public $deleted = array();
-    public $committed = false;
-
-    public function reset()
-    {
-        $this->added = array();
-        $this->deleted = array();
-        $this->committed = false;
-    }
-
-    public function add($object)
-    {
-        $res = array();
-
-        $res['ID'] = $object->ID;
-
-        foreach ($this->getFieldsIterator() as $name => $field) {
-            $val = $this->_getFieldValue($object, $field);
-            $res[$name] = $val;
-        }
-
-        $this->added[] = $res;
-    }
-
-    public function getAdded($fields = array())
-    {
-        $res = array();
-
-        foreach ($this->added as $added) {
-            $filtered = array();
-            foreach ($fields as $field) {
-                if (isset($added[$field])) {
-                    $filtered[$field] = $added[$field];
-                }
-            }
-            $res[] = $filtered;
-        }
-
-        return $res;
-    }
-
-    public function delete($base, $id, $state)
-    {
-        $this->deleted[] = array('base' => $base, 'id' => $id, 'state' => $state);
-    }
-
-    public function commit()
-    {
-        $this->committed = true;
-    }
-
-    public function getIndexName()
-    {
-        return get_class($this);
-    }
-
-    public function getIsCommitted()
-    {
-        return $this->committed;
-    }
-
-    public function getService()
-    {
-        // Causes commits to the service to be redirected back to the same object
-        return $this;
-    }
 }
